+++ colossalai/inference/core/llm_engine.py
@@ -30,7 +30,7 @@ from colossalai.lazy import LazyInitContext
 from colossalai.logging import get_dist_logger
 from colossalai.shardformer.policies.base_policy import Policy
 
-from .base_engine import InferenceEngine
+from .base_engine import BaseEngine
 from .request_handler import RequestHandler
 
 PP_AXIS, TP_AXIS = 0, 1
@@ -43,7 +43,7 @@ _supported_models = {
 _BATCH_SIZES_TO_CAPTURE = [1, 2, 4] + [8 * i for i in range(1, 33)]
 
 
-class LLMEngine(InferenceEngine):
+class LLMEngine(BaseEngine):
 
     """
     InferenceEngine which manages the inference process..
